[enocean] Humidity and temperature readings for EEP A5_10
* Added I8N for new props
* Removed some magic numbers

fixes #15458

+++ bundles/org.openhab.binding.enocean/src/main/java/org/openhab/binding/enocean/internal/eep/A5_10/A5_10.java
@@ -46,8 +46,8 @@ public abstract class A5_10 extends _4BSMessage {
         return getDB2Value();
     }
 
-    protected int getMaxUnscaledValue() {
-        return 255;
+    protected double getMaxUnscaledValue() {
+        return 255.0;
     }
 
     protected double getTempScalingFactor() {
