+++ bundles/org.openhab.binding.enocean/src/main/java/org/openhab/binding/enocean/internal/eep/A5_10/A5_10_18.java
@@ -29,8 +29,8 @@ public class A5_10_18 extends A5_10 {
     }
 
     @Override
-    protected int getMaxUnscaledValue() {
-        return 250;
+    protected double getMaxUnscaledValue() {
+        return 250.0;
     }
 
     @Override
